+++ packages/create-block/lib/init-block.js
@@ -15,7 +15,7 @@ const { writeOutputTemplate } = require( './output' );
 async function initBlockJSON( {
 	$schema,
 	apiVersion,
-	blockOnly,
+	plugin,
 	slug,
 	namespace,
 	title,
@@ -34,8 +34,8 @@ async function initBlockJSON( {
 	info( '' );
 	info( 'Creating a "block.json" file.' );
 
-	const outputFile = blockOnly
-		? join( process.cwd(), folderName, slug, 'block.json' )
+	const outputFile = ! plugin
+		? join( process.cwd(), slug, 'block.json' )
 		: join( process.cwd(), slug, folderName, 'block.json' );
 	await makeDir( dirname( outputFile ) );
 	await writeFile(
@@ -69,7 +69,7 @@ async function initBlockJSON( {
 module.exports = async function ( outputTemplates, view ) {
 	await Promise.all(
 		Object.keys( outputTemplates ).map( async ( outputFile ) => {
-			const pathName = view.blockOnly
+			const pathName = ! view.plugin
 				? join( process.cwd(), view.slug, outputFile )
 				: join( view.folderName, outputFile );